refactor: rename <HomePageBanner/> -> <CurricHomePageBanner/>
