+++ src/components/CurriculumComponents/Banners/Banners.tsx
@@ -1,8 +1,8 @@
-import HomePageBanner from "@/components/CurriculumComponents/HomePageBanner";
+import CurricHomePageBanner from "@/components/CurriculumComponents/CurricHomePageBanner";
 
 export default function Banners() {
   return (
-    <HomePageBanner
+    <CurricHomePageBanner
       background="lemon"
       newText="Subjects added"
       ctaText="See curriculum plans"
