+++ src/components/CurriculumComponents/CurricHomePageBanner/CurricHomePageBanner.tsx
