+++ src/components/CurriculumComponents/CurricHomePageBanner/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./CurricHomePageBanner";
