--- src/components/CurriculumComponents/HomePageBanner/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./HomePageBanner";
